Cleanup of code moved from TypeInferrer to Field subclasses
