+++ src/main/java/io/frictionlessdata/tableschema/field/YearMonthField.java
@@ -3,13 +3,20 @@ package io.frictionlessdata.tableschema.field;
 import io.frictionlessdata.tableschema.TypeInferrer;
 import io.frictionlessdata.tableschema.exceptions.ConstraintsException;
 import io.frictionlessdata.tableschema.exceptions.InvalidCastException;
+import io.frictionlessdata.tableschema.exceptions.TypeInferringException;
 import org.joda.time.DateTime;
+import org.joda.time.format.DateTimeFormat;
+import org.joda.time.format.DateTimeFormatter;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
 import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class YearMonthField extends Field<DateTime> {
+    // yyyy-MM
+    private static final String REGEX_YEARMONTH = "([0-9]{4})-(1[0-2]|0[1-9])";
 
     public YearMonthField(String name) {
         super(name, FIELD_TYPE_YEARMONTH);
@@ -26,6 +33,17 @@ public class YearMonthField extends Field<DateTime> {
 
     @Override
     DateTime getCastValue(String value, String format, Map<String, Object> options) throws InvalidCastException, ConstraintsException {
-        return TypeInferrer.getInstance().castYearmonth(value, format, options);
+        Pattern pattern = Pattern.compile(REGEX_YEARMONTH);
+        Matcher matcher = pattern.matcher(value);
+
+        if(matcher.matches()){
+            DateTimeFormatter formatter = DateTimeFormat.forPattern("yyyy-MM");
+            DateTime dt = formatter.parseDateTime(value);
+
+            return dt;
+
+        }else{
+            throw new TypeInferringException();
+        }
     }
 }
